Gutenboarding: visual fixes for #40569
